rename scheduling modal
