+++ packages/frinx-workflow-ui/src/App.js
@@ -14,7 +14,7 @@ import bulkReducer from './store/reducers/bulk';
 import searchReducer from './store/reducers/searchExecs';
 import WorkflowDefinitions from './pages/workflow-list/WorkflowDefs/workflow-definitions';
 import WorkflowExec from './pages/workflow-list/workflow-execution/workflow-exec';
-import Scheduling from './pages/workflow-list/Scheduling/Scheduling';
+import Scheduling from './pages/workflow-list/scheduling/scheduling';
 import EventListeners from './pages/workflow-list/EventListeners/event-listeners';
 import TaskList from './pages/workflow-list/Tasks/task-list';
 import PollData from './pages/workflow-list/PollData/poll-data';
